multipart bridge
